+++ apps/ui/lib/ui_web/graph/types/event_athlete.ex
@@ -6,9 +6,9 @@ defmodule UIWeb.Graph.Schema.Types.EventAthlete do
     @desc "id for event athlete"
     field :id, non_null(:id)
     @desc "How much the athlete costs in the event"
-    field :cost, non_null(:integer)
+    field :cost, non_null(:decimal)
     @desc "How much the athlete won in the event"
-    field :winnings, :integer
+    field :winnings, :decimal
     @desc "The athletes information"
     field :athlete, non_null(:athlete) do
       resolve &Athlete.by_id/3
